Finalised ReadME Moved all files into folder with project name and student number
