+++ .gitignore
@@ -1,4 +1,4 @@
-/datasets/
-**/__pycache__/
-/*_generated/
-/ModelParams/
+OASIS_VQ_VAE_s4698512/datasets/
+OASIS_VQ_VAE_s4698512/**/__pycache__/
+OASIS_VQ_VAE_s4698512/*_generated/
+OASIS_VQ_VAE_s4698512/ModelParams/
