--- Checklist.md
@@ -1,34 +1,0 @@
-Checklist
-
-1. Forked Pattern Analysis library on Github. Done
-2. Checked out the topic-recognition branch within your fork.
-
-Made branch called topic-recognition and have checked it out
-
-3. Choose a problem from section 1
-
-Chosen problem: Number 8. Generative model of OASIS brain, ADNI brain or OAI AKOA knee data set.
-
-Will choose ADNI brain, though I may go with OASIS too.
-
-Created folder called BrainVQVAE_s4698512 in root directory of repository.
-
-4. Create README.md file whic describes report. Done. Have renamed old README.md to README_from_repo.md
-
-5. Complete solution to the recognition problem by building your model. In progress.
-
-Ensure to make git commits along the way
-
-6. Do not commit datasets.
-   Have excludes datasets folder in .gitignore file
-
-7. Write a test drivers script that calls and runs your algorithm
-   In progress
-
-8. Complete documentation of your algorithm including a summary of your results and model in your README.md file.
-
-9. Lodge a pull request
-   In progress
-
-10. Complete turn it in submission of the PDF and of the README.md file
-    In progress
